fix(structure): reset historyParams only if selectedPerspective changes
