+++ packages/sanity/src/structure/panes/document/DocumentEventsPane.tsx
@@ -1,4 +1,4 @@
-import {useEffect, useMemo, useRef} from 'react'
+import {useMemo} from 'react'
 import {
   EventsProvider,
   getDraftId,
@@ -8,7 +8,6 @@ import {
   usePerspective,
   useReleases,
 } from 'sanity'
-import {useEffectEvent} from 'use-effect-event'
 
 import {usePaneRouter} from '../../components'
 import {EMPTY_PARAMS} from './constants'
@@ -17,7 +16,7 @@ import {DocumentPaneProvider} from './DocumentPaneProvider'
 import {type DocumentPaneProviderProps} from './types'
 
 export const DocumentEventsPane = (props: DocumentPaneProviderProps) => {
-  const {params = EMPTY_PARAMS, setParams} = usePaneRouter()
+  const {params = EMPTY_PARAMS} = usePaneRouter()
   const options = usePaneOptions(props.pane.options, params)
 
   const {selectedPerspectiveName} = usePerspective()
@@ -40,29 +39,6 @@ export const DocumentEventsPane = (props: DocumentPaneProviderProps) => {
     }
     return options.id
   }, [archivedReleases, historyVersion, selectedPerspectiveName, options.id])
-
-  const isMounted = useRef(false)
-
-  const updateHistoryParams = useEffectEvent((_perspective?: string) => {
-    setParams({
-      ...params,
-      // Reset the history related params when the perspective changes, as they don't make sense
-      // in the context of the new perspective - preserveRev is used when setting draft revision.
-      rev: params.preserveRev === 'true' ? params.rev : undefined,
-      preserveRev: undefined,
-      since: undefined,
-      historyVersion: undefined,
-    })
-  })
-  useEffect(() => {
-    // Skip the first run to avoid resetting the params on initial load
-    if (isMounted.current) {
-      updateHistoryParams(selectedPerspectiveName)
-    } else {
-      isMounted.current = true
-    }
-    // TODO: Remove `updateHistoryParams` as a dependency when react eslint plugin is updated
-  }, [selectedPerspectiveName, updateHistoryParams])
 
   const eventsStore = useEventsStore({documentId, documentType: options.type, rev, since})
